Extract ContextUserIsActiveAsync to ContextService
