+++ Domain/Interfaces/IHttpContextService.cs
@@ -1,0 +1,15 @@
+using Data.Models.Dbo;
+using Domain.Models.State;
+
+namespace Domain.Interfaces;
+
+public interface IHttpContextService
+{
+    /// <summary>
+    /// Is the context user active, authenticated, not blocked etc.
+    /// Essentially, free to make actions on the app (create clubs, comments, etc.)
+    /// </summary>
+    /// <param name="userId"></param>
+    /// <returns></returns>
+    Task<ResultState<ApplicationUserDbo?>> ContextUserIsActiveAsync();
+}
